src/end_to_end/edge_device/video_player/requirements.txt
opencv-python
paho_mqtt